Extract `ActionSeedSwipeConfiguration` to new file
