+++ Mlem/App/Enums/Interaction/CommunityActionConfiguration.swift
@@ -67,21 +67,3 @@ struct CommunityActionConfiguration: Codable {
         try container.encode(self.swipes_, forKey: .swipes)
     }
 }
-
-struct ActionSeedSwipeConfiguration: Encodable {
-    var leading: [ActionSeed]   
-    var trailing: [ActionSeed]
-
-    enum CodingKeys: CodingKey {
-        case leading, trailing
-    }
-}
-
-extension ActionSeedSwipeConfiguration {
-    init(from container: KeyedDecodingContainer<CodingKeys>, availableActions: [ActionSeed]) throws {
-        let leading = try container.decode([String].self, forKey: .leading) 
-        self.leading = leading.compactMap { key in availableActions.first(where: {$0.key == key}) }
-        let trailing = try container.decode([String].self, forKey: .trailing) 
-        self.trailing = trailing.compactMap { key in availableActions.first(where: {$0.key == key}) }
-    }
-}
